+++ protocol_wire/include/protocol_wire/SellerTerms.hpp
@@ -31,7 +31,7 @@ namespace protocol_wire {
 
         SellerTerms();
 
-        SellerTerms(quint64 minPrice, quint32 minLock, quint32 maxSellers, quint64 minContractFeePerKb, quint64 settlementFee);
+        SellerTerms(quint64 minPrice, quint16 minLock, quint32 maxSellers, quint64 minContractFeePerKb, quint64 settlementFee);
 
         bool operator==(const SellerTerms & rhs) const;
 
@@ -54,8 +54,8 @@ namespace protocol_wire {
         quint64 minPrice() const;
         void setMinPrice(quint64 minPrice);
 
-        quint32 minLock() const;
-        void setMinLock(quint32 minLock);
+        quint16 minLock() const;
+        void setMinLock(quint16 minLock);
 
         quint32 maxSellers() const;
         void setMaxSellers(quint32 maxSellers);
@@ -72,9 +72,8 @@ namespace protocol_wire {
         //int64_t _price;
         quint64 _minPrice;
 
-        // When refund is spendable at the earliest (***UNITS?***)
-        //uint32_t _lock;
-        quint32 _minLock;
+        // When refund is spendable at the earliest
+        quint16 _minLock;
 
         // Maximum number of sellers accepted in contract
         //uint32_t _maxSellers;
